+++ src/main/webapp/ui/src/tinyMCE/stoichiometry/__tests__/StoichiometryDialog.spec.tsx
