+++ src/main/webapp/ui/src/tinyMCE/stoichiometry/__tests__/StoichiometryDialog.story.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { StyledEngineProvider, ThemeProvider } from "@mui/material/styles";
-import createAccentedTheme from "../../accentedTheme";
-import { ACCENT_COLOR } from "../../assets/branding/chemistry";
-import StoichiometryDialog from "./StoichiometryDialog";
-import Alerts from "../../components/Alerts/Alerts";
+import createAccentedTheme from "../../../accentedTheme";
+import { ACCENT_COLOR } from "../../../assets/branding/chemistry";
+import StoichiometryDialog from "../StoichiometryDialog";
+import Alerts from "../../../components/Alerts/Alerts";
 import { createStoichiometryTheme } from "@/tinyMCE/stoichiometry/theme";
 
 export function StoichiometryDialogWithCalculateButtonStory({
