+++ src/main/webapp/ui/src/tinyMCE/stoichiometry/__tests__/StoichiometryTable.spec.tsx
