+++ src/main/webapp/ui/src/tinyMCE/stoichiometry/__tests__/StoichiometryTable.story.tsx
@@ -4,9 +4,9 @@ import Box from "@mui/material/Box";
 import Typography from "@mui/material/Typography";
 import CircularProgress from "@mui/material/CircularProgress";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import createAccentedTheme from "../../accentedTheme";
-import { ACCENT_COLOR } from "../../assets/branding/chemistry";
-import Alerts from "../../components/Alerts/Alerts";
+import createAccentedTheme from "../../../accentedTheme";
+import { ACCENT_COLOR } from "../../../assets/branding/chemistry";
+import Alerts from "../../../components/Alerts/Alerts";
 import StoichiometryTable from "@/tinyMCE/stoichiometry/StoichiometryTable";
 import { StoichiometryTableControllerProvider } from "@/tinyMCE/stoichiometry/StoichiometryTableControllerContext";
 import { createStoichiometryTheme } from "@/tinyMCE/stoichiometry/theme";
