+++ src/main/webapp/ui/src/tinyMCE/stoichiometry/__tests__/StoichiometryTableInventoryLinkCell.test.tsx
