+++ src/main/webapp/ui/src/tinyMCE/stoichiometry/__tests__/utils.test.ts
